# AsynCUDA13.Client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AsynCUDA13.API v1",
    "version": "v1"
  },
  "paths": {
...
    "/api/RuntimeFourier/CuFFT": {
      "post": {
        "tags": [
          "RuntimeFourier"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeFourierRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeFourierRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeFourierRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeFourierResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeFourierResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeFourierResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/RuntimeFourier/CuFFT/audio": {
      "post": {
        "tags": [
          "RuntimeFourier"
        ],
        "parameters": [
          {
            "name": "chunkSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "overlap",
            "in": "query",
            "schema": {
              "type": "number",
              "format": "float",
              "default": 0.5
            }
          },
          {
            "name": "autoPullResult",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          },
          {
            "name": "keepDataOrBuffer",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            },
            "text/json": {
              "schema": {
                "type": "string"
              }
            },
            "application/*+json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeFourierResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeFourierResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeFourierResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ISimdPayload": {
        "type": "object",
        "properties": {
          "elementType": {
            "type": "string",
            "nullable": true
          },
          "asyncCall": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "RuntimeFourierRequest": {
        "required": [
          "memoryInfo"
        ],
        "type": "object",
        "properties": {
          "memoryInfo": {
            "$ref": "#/components/schemas/RuntimeMemInfo"
          },
          "inverse": {
            "type": "boolean",
            "nullable": true
          },
          "asyncCall": {
            "type": "boolean"
          },
          "keepInputBuffer": {
            "type": "boolean"
          },
          "autoPullResult": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RuntimeFourierResponse": {
        "required": [
          "inputMemoryInfoReference"
        ],
        "type": "object",
        "properties": {
          "inputMemoryInfoReference": {
            "$ref": "#/components/schemas/RuntimeMemInfo"
          },
          "outputMemoryInfoReference": {
            "$ref": "#/components/schemas/RuntimeMemInfo"
          },
          "indexPointer": {
            "type": "string",
            "nullable": true
          },
          "outputPayload": {
            "$ref": "#/components/schemas/ISimdPayload"
          },
          "elapsedMs": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RuntimeMemInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "indexPointer": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "elementType": {
            "type": "string",
            "nullable": true
          },
          "indexLength": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "elementSize": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "nullable": true,
            "readOnly": true
          },
          "longCount": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "totalLength": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "totalSize": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "lengths": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "pointers": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isValid": {
            "type": "boolean",
            "readOnly": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}